+++ src/App-Services/Create-Web-App-with-App-Service-Plan-Linux.ps1
@@ -13,7 +13,6 @@ param (
     [Parameter(Mandatory)][string] $AppServiceResourceGroupName,
     [Parameter(Mandatory)][string] $AppServiceName,
     [Parameter(Mandatory, ParameterSetName = 'default')][Parameter(Mandatory, ParameterSetName = 'DeploymentSlot')][string] $AppServiceRunTime,
-    [Parameter(Mandatory)][string] $AppServiceDiagnosticsName,
     [Alias("LogAnalyticsWorkspaceName")]
     [Parameter(Mandatory)][string] $LogAnalyticsWorkspaceResourceId,
     [Parameter(Mandatory)][string] $DNSZoneResourceGroupName,
@@ -36,10 +35,12 @@ Import-Module "$PSScriptRoot\..\AzDocs.Common" -Force
 
 Write-Header -ScopedPSCmdlet $PSCmdlet
 
+Write-Warning 'This script is deprecated. Please use the Create-Web-App.ps1 and the Create-App-Service-Plan-Linux.ps1 instead.'
+
 # Create AppServicePlan
 & "$PSScriptRoot\Create-App-Service-Plan-Linux.ps1" -AppServicePlanName $AppServicePlanName -AppServicePlanResourceGroupName $AppServicePlanResourceGroupName -AppServicePlanSkuName $AppServicePlanSkuName -AppServicePlanNumberOfWorkerInstances $AppServicePlanNumberOfWorkerInstances -ResourceTags ${ResourceTags}
 
 # Create AppService
-& "$PSScriptRoot\Create-Web-App-Linux.ps1" @PSBoundParameters
+& "$PSScriptRoot\Create-Web-App.ps1" @PSBoundParameters
 
 Write-Footer -ScopedPSCmdlet $PSCmdlet