+++ src/Storage-Accounts/Create-Fileshare-in-StorageAccount.ps1
@@ -5,7 +5,9 @@ param (
     [Alias("FileshareStorageAccountName")]
     [Parameter(Mandatory)][string] $StorageAccountName,
     [Alias("ShareName")]
-    [Parameter(Mandatory)][string] $FileshareName
+    [Parameter(Mandatory)][string] $FileshareName,
+    # [Parameter(Mandatory)][string] $FileshareDiagnosticsName,
+    [Parameter(Mandatory)][string] $LogAnalyticsWorkspaceResourceId
 )
 
 #region ===BEGIN IMPORTS===
@@ -14,6 +16,11 @@ Import-Module "$PSScriptRoot\..\AzDocs.Common" -Force
 
 Write-Header -ScopedPSCmdlet $PSCmdlet
 
+# Create storage fileshare
 Invoke-Executable az storage share-rm create --storage-account $StorageAccountName --name $FileshareName
+
+# Enable diagnostic settings for storage fileshare
+$storageAccountId = (Invoke-Executable az storage account show --name $StorageAccountName --resource-group $StorageAccountResourceGroupName | ConvertFrom-Json).id 
+Set-DiagnosticSettings -ResourceId "$storageAccountId/fileServices/default" -ResourceName $StorageAccountName -LogAnalyticsWorkspaceResourceId $LogAnalyticsWorkspaceResourceId -Logs "[{ 'category': 'StorageRead', 'enabled': true }, { 'category': 'StorageWrite', 'enabled': true }, { 'category': 'StorageDelete', 'enabled': true }]".Replace("'", '\"')  -Metrics "[ { 'category': 'AllMetrics', 'enabled': true } ]".Replace("'", '\"') 
 
 Write-Footer -ScopedPSCmdlet $PSCmdlet