+++ Wiki/Azure/CHANGELOG.md
@@ -2,24 +2,76 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2021.07.26]
+
+New scripts and improvements!
+We've improved the creation of the diagnostic settings which are crucial for the metrics and logs of your resources! Next to that, we've added some new features
+to use in your platform.
+
+Happy deploying!
+
+### Breaking Changes
+
+- Removed `AppConfigDiagnosticsName` from `Create-App-Configuration.ps1`
+- Removed `AppServiceDiagnosticsName` from `Create-Web-App-Linux.ps1`
+- Removed `AppServiceDiagnosticsName` from `Create-Web-App-Windows.ps1`
+- Removed `AppServiceDiagnosticsName` from `Create-Web-App-with-App-Service-Plan-Windows.ps1`
+- Removed `AppServiceDiagnosticsName` from `Create-Web-App-with-App-Service-Plan-Linux.ps1`
+- Removed `FunctionAppDiagnosticsName` from `Create-Function-App-Linux.ps1`
+- Removed `FunctionAppDiagnosticsName` from `Create-Function-App-Windows.ps1`
+- Removed `FunctionAppDiagnosticsName` from `Create-Function-App-Windows-with-App-Service-Plan.ps1`
+- Removed `FunctionAppDiagnosticsName` from `Create-Function-App-Linux-with-App-Service-Plan.ps1`
+- Removed `KeyvaultDiagnosticsName` from `Create-Keyvault.ps1`
+- Added diagnostic settings to `Create-Blobcontainer-in-StorageAccount.ps1`
+  - Added mandatory parameters `StorageAccountResourceGroupName` and `LogAnalyticsWorkspaceResourceId`
+- Added diagnostic settings to `Create-Fileshare-in-StorageAccount.ps1`
+  - Added mandatory parameters `LogAnalyticsWorkspaceResourceId`
+- Added diagnostic settings to `Create-Queue-in-StorageAccount.ps1`
+  - Added mandatory parameters `StorageAccountResourceGroupName` and `LogAnalyticsWorkspaceResourceId`
+- Added diagnostic settings to `Create-Storage-Account.ps1`
+  - Added mandatory parameters `LogAnalyticsWorkspaceResourceId`
+- Added diagnostic settings to `Create-Container-Registry.ps1`
+  - Added mandatory parameters `LogAnalyticsWorkspaceResourceId`
+- Added diagnostic settings to `Create-MySQL-Server.ps1`
+  - Added mandatory parameters `LogAnalyticsWorkspaceResourceId`
+- Added diagnostic settings to `Create-PostgreSQL-Server.ps1`
+  - Added mandatory parameters `LogAnalyticsWorkspaceResourceId`
+- Added diagnostic settings to `Create-RedisCache-Instance.ps1`
+  - Added mandatory parameters `LogAnalyticsWorkspaceResourceId`
+- Added diagnostic settings to `Create-SQL-Database.ps1`
+  - Added mandatory parameters `LogAnalyticsWorkspaceResourceId`
+- Added diagnostic settings to `Create-AppInsights-Resource.ps1`
+  - Added mandatory parameters `DiagnosticSettingLogAnalyticsWorkspaceResourceId`
+- Added diagnostic settings to `Create-Application-Gateway.ps1`
+  - Added mandatory parameters `LogAnalyticsWorkspaceResourceId`
+
+### Added
+
+- Added the possibility to add an elastic pool to your sql server in script `Add-Elastic-Pool-To-SQL-Server.ps1`
+- Added the possibility to add your database to an elastic pool in script `Create-SQL-Database.ps1`
+- Added the possibility to deprovision an entry point in the Application Gateway (including certificates) in script `Remove-Application-Gateway-Entrypoint.ps1`
+- Added the possiblity to add a custom dns server to a VNET in script `Add-Custom-DNS-Server-To-VNET.ps1`
+
 ## [2021.07.19]
+
 We are excited to bring you the big release of the overall documentation containing architectural explanations!
 
 ### Breaking Changes
+
 None!
 
 ### Added
 
 - Documentation
-    - Recommended architecture
-    - Which components are available?
-    - Tutorials on how to setup this boilerplate in your Azure DevOps environment.
-    - How to use these scripts?
-    - Guidelines for creating new scripts & contributing to this repository.
-    - Application Gateway information & Best practises
-    - Networking information (private endpoints & vnet whitelisting)
-    - Logging & monitoring
-    - Deprovisioning
+  - Recommended architecture
+  - Which components are available?
+  - Tutorials on how to setup this boilerplate in your Azure DevOps environment.
+  - How to use these scripts?
+  - Guidelines for creating new scripts & contributing to this repository.
+  - Application Gateway information & Best practises
+  - Networking information (private endpoints & vnet whitelisting)
+  - Logging & monitoring
+  - Deprovisioning
 - Added WAF options to the [Create Application Gateway](/Azure/Azure-CLI-Snippets/Application-Gateway/Create-Application-Gateway) script.
 - Added the [Add DDoS Protection To Vnet](/Azure/Azure-CLI-Snippets/Networking/Add-DDoS-Protection-To-Vnet) script.
 - Reset the Wiki order of components to alphabetical (for easier finding your components).
@@ -27,6 +79,7 @@ None!
 Note: as always this is work in progress and will be subjected to changes & additions.
 
 ## [2021.06.23]
+
 A networking release! We've made sure it is easier to interact with different kind of networking whitelisting/dewhitelisting actions within you platform.
 
 ### Breaking Changes